rafiki/utils/requirements.txt
certifi==2018.8.24
chardet==3.0.4
idna==2.7
numpy==1.14.5
Pillow==5.3.0
PyJWT==1.6.4
requests==2.19.1
urllib3==1.23
